Refactor connections
