--- src/components/Window/NodeConnections.module.scss
@@ -1,34 +1,0 @@
-.incomingWrapper {
-  position: absolute;
-  z-index: 10;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  gap: 10px;
-}
-
-.nodeWrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: 15px;
-  width: fit-content;
-  background-color: rgba(0, 0, 0, 0.461);
-  padding: 5px;
-  border-radius: 5px;
-  opacity: 0.5;
-  transition: opacity 0.2s ease-in-out;
-  &:hover {
-    opacity: 1;
-  }
-  &.hidden {
-    opacity: 0;
-  }
-}
-
-.node {
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-}
