+++ src/ui/Connections/NodeConnector.module.scss
@@ -1,4 +1,4 @@
-.incomingWrapper {
+.wrapper {
   position: absolute;
   z-index: 10;
   display: flex;
